Update Svelte project layouts, migrate to static adapter and add firebase dependency
The code changes reflect updates to Svelte's project layouts by adding text to headers in the main and hardware sections. The switch from the auto adapter to the static adapter in the Svelte config file has been made for better control over building configurations. Lastly, a new dependency has been added to the package.json for compatibility with Firebase.